[Refactor] 폴더명 변경, 타입 공통화
