+++ src/components/Gnb/HeaderBebridge.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { useRouter } from "next/router";
-import SkeletonUser from "./skeletonUser";
-import { MemberType, UserType } from "./type";
+import SkeletonUser from "../../shared/skeletonUser";
+import { MemberType, UserType } from "../../types/users";
 import { getMembers } from "@/api/members";
 import { getUserInfo } from "@/api/user";
 import RandomProfile from "../table/member/RandomProfile";
